// frontend/src/sections/rnr-forms/list/create/createRnrForm.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../../api/graphql";

export type RnrProgramsVariables = {
  storeId: string;
};

export type RnrProgramsResult = {
  programs: ({
  nodes: Array<{
  id: string;
  name: string;
  isImmunisation: boolean;
}>;
});
};

export const RnrPrograms = {
  query: "query RnrPrograms($storeId: String!) {\n  programs(storeId: $storeId, sort: {key: name}) {\n    ... on ProgramConnector {\n      nodes {\n        id\n        name\n        isImmunisation\n      }\n    }\n  }\n}",
} as TypedDocument<RnrProgramsResult, RnrProgramsVariables>;

export type RnrSchedulesVariables = {
  storeId: string;
  programId: string;
};

export type RnrSchedulesResult = {
  schedulesWithPeriodsByProgram: ({
  nodes: Array<{
  id: string;
  name: string;
  periods: Array<{
  id: string;
  inUse: boolean;
  period: {
  id: string;
  name: string;
  startDate: string;
  endDate: string;
};
}>;
}>;
});
};

export const RnrSchedules = {
  query: "query RnrSchedules($storeId: String!, $programId: String!) {\n  schedulesWithPeriodsByProgram(storeId: $storeId, programId: $programId) {\n    ... on PeriodSchedulesConnector {\n      nodes {\n        id\n        name\n        periods {\n          id\n          inUse\n          period {\n            id\n            name\n            startDate\n            endDate\n          }\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<RnrSchedulesResult, RnrSchedulesVariables>;

export type InsertRnrFormVariables = {
  storeId: string;
  input: {
    id: string;
    supplierId: string;
    programId: string;
    periodId: string;
  };
};

export type InsertRnrFormResult = {
  insertRnrForm: ({
  id: string;
});
};

export const InsertRnrForm = {
  query: "mutation InsertRnrForm($storeId: String!, $input: InsertRnRFormInput!) {\n  insertRnrForm(storeId: $storeId, input: $input) {\n    ... on RnRFormNode {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertRnrFormResult, InsertRnrFormVariables>;